requirements.txt
torch
torchvision
torchaudio
flask==2.3.2
flask-cors==4.0.0
python-multipart==0.0.6
librosa==0.10.0
joblib==1.3.1
numpy==1.24.3
h5py==3.11.0
scikit-learn==1.6.1
fer==22.4.3
opencv-python==4.7.0.72
pillow==9.5.0
PyJWT==2.8.0